refactor: restructure datasets endpoints to use files for different datasets endpoint entities
