+++ tests/unit/server/api/v1/datasets/test_create_dataset_vector_settings.py
@@ -15,7 +15,7 @@
 from uuid import UUID
 
 import pytest
-from argilla.server.apis.v1.handlers.datasets import CREATE_DATASET_VECTOR_SETTINGS_MAX_COUNT
+from argilla.server.apis.v1.handlers.datasets.datasets import CREATE_DATASET_VECTOR_SETTINGS_MAX_COUNT
 from httpx import AsyncClient
 
 from tests.factories import DatasetFactory, VectorSettingsFactory
