+++ tests/unit/server/api/v1/datasets/test_search_dataset_records.py
@@ -16,7 +16,7 @@ from uuid import UUID, uuid4
 
 import pytest
 from argilla._constants import API_KEY_HEADER_NAME
-from argilla.server.apis.v1.handlers.datasets import LIST_DATASET_RECORDS_LIMIT_LE
+from argilla.server.apis.v1.handlers.datasets.records import LIST_DATASET_RECORDS_LIMIT_LE
 from argilla.server.enums import RecordInclude
 from argilla.server.search_engine import SearchEngine, SearchResponseItem, SearchResponses
 from httpx import AsyncClient
