+++ tests/unit/server/api/v1/test_list_dataset_records.py
@@ -17,7 +17,7 @@ from uuid import uuid4
 
 import pytest
 from argilla._constants import API_KEY_HEADER_NAME
-from argilla.server.apis.v1.handlers.datasets import LIST_DATASET_RECORDS_LIMIT_DEFAULT
+from argilla.server.apis.v1.handlers.datasets.records import LIST_DATASET_RECORDS_LIMIT_DEFAULT
 from argilla.server.enums import RecordInclude, RecordSortField, ResponseStatus, UserRole
 from argilla.server.models import Dataset, Question, Record, Response, Suggestion, User, Workspace
 from argilla.server.search_engine import (
